join registration process into one service
