+++ src/main/java/ita/softserve/course_evaluation/registration/EmailValidator.java
@@ -8,7 +8,7 @@ import java.util.function.Predicate;
 public class EmailValidator implements Predicate<String> {
     @Override
     public boolean test(String s) {
-        //TODO: Regex email
+        //TODO: Implement Regex email
         return true;
     }
 }
