+++ src/main/java/ita/softserve/course_evaluation/service/UserService.java
@@ -17,6 +17,4 @@ public interface UserService {
 
 	void updatePassword(UpdatePasswordDto updatePasswordDto, String email);
 
-	String signUp(User user);
-	int enableAppUser(String email);
 }
