+++ src/main/java/ita/softserve/course_evaluation/service/impl/AuthServiceImpl.java
@@ -5,6 +5,7 @@ import ita.softserve.course_evaluation.dto.SimpleUserDto;
 import ita.softserve.course_evaluation.dto.SimpleUserDtoResponseMapper;
 import ita.softserve.course_evaluation.entity.User;
 import ita.softserve.course_evaluation.exception.EmailNotConfirmedException;
+import ita.softserve.course_evaluation.registration.RegistrationServiceImpl;
 import ita.softserve.course_evaluation.repository.UserRepository;
 import ita.softserve.course_evaluation.security.jwt.JwtTokenProvider;
 import ita.softserve.course_evaluation.service.AuthService;
@@ -28,12 +29,14 @@ public class AuthServiceImpl implements AuthService {
 	private final PasswordEncoder passwordEncoder;
 	private final UserRepository userRepository;
 	private final JwtTokenProvider jwtTokenProvider;
+	private final RegistrationServiceImpl registrationService;
 	
-	public AuthServiceImpl(AuthenticationManager authenticationManager, PasswordEncoder passwordEncoder, UserRepository userRepository, JwtTokenProvider jwtTokenProvider) {
+	public AuthServiceImpl(AuthenticationManager authenticationManager, PasswordEncoder passwordEncoder, UserRepository userRepository, JwtTokenProvider jwtTokenProvider, RegistrationServiceImpl registrationService) {
 		this.authenticationManager = authenticationManager;
 		this.passwordEncoder = passwordEncoder;
 		this.userRepository = userRepository;
 		this.jwtTokenProvider = jwtTokenProvider;
+		this.registrationService = registrationService;
 	}
 	
 	public ResponseEntity<?> getLoginCredentials(AuthenticateRequestDto request) {
